+++ jsr223/src/main/resources/META-INF/services/javax.script.ScriptEngineFactory
@@ -1,1 +1,1 @@
-party.iroiro.luajava.LuaScriptEngineFactory
+party.iroiro.luajava.jsr223.LuaScriptEngineFactory
